requirements.txt
pytest
pytest-cov
pytest-mock
mock
pytest-ordering
mysql-connector
psycopg2
pymongo
flask
setuptools
wheel
connexion
python-jose
makefun
python-dateutil